joycemay-linebot/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
line-bot-sdk==1.18.0